+++ src/views/SearchView/SearchView.js
@@ -72,7 +72,7 @@ const SearchView = () => {
     return (
       <span>
         {intl.formatMessage({
-          id: 'stripes-authority-components.search-results-list.paneSub',
+          id: 'ui-plugin-find-authority.search-results-list.paneSub',
         }, {
           totalRecords,
         })}