+++ CHANGELOG.md
@@ -1,3 +1,4 @@
 ## [0.1.0] (IN PROGRESS)
 
 * Create an empty modal window for find Authorities plugin. Refs UIPFAUTH-7
+* Select a MARC authority record modal > Results list - Search. Refs UIPFAUTH-3