+++ CHANGELOG.md
@@ -1,4 +1,6 @@
-## [0.1.0] (IN PROGRESS)
+# Change history for ui-plugin-find-authority
+
+## [1.0.0] (IN PROGRESS)
 
 * Create an empty modal window for find Authorities plugin. Refs UIPFAUTH-7
 * Select a MARC authority record modal > Results list - Search. Refs UIPFAUTH-3